+++ plugins/BuffAutoAcceptOffer.py
@@ -4,11 +4,11 @@ import pickle
 import time
 
 import apprise
+import Apprise
 import json5
 import requests
 from apprise.AppriseAsset import AppriseAsset
 from requests.exceptions import ProxyError
-from Apprise.server_chan import *
 
 from steampy.exceptions import ConfirmationExpected, InvalidCredentials
 from utils.buff_helper import get_valid_session_for_buff
