+++ plugins/BuffAutoOnSale.py
@@ -6,10 +6,10 @@ import time
 from typing import Any
 
 import apprise
+import Apprise
 import json5
 import requests
 from apprise.AppriseAsset import AppriseAsset
-from Apprise.server_chan import *
 from bs4 import BeautifulSoup
 from requests.exceptions import ProxyError
 from steampy.exceptions import InvalidCredentials
